+++ src/reactive/operator.js
@@ -1,0 +1,7 @@
+export default class Operator {
+  constructor() {
+    this.filter = predicate => { /* to be implemented */ };
+    this.map = predicate => { /* to be implemented */ };
+    this.reduce = predicate => { /* to be implemented */ };
+  }
+}